Restructured file layout.
